examples: freestanding per-example layout (config/app/platform/generated) (#10)
* examples: freestanding per-example layout (config/app/platform/generated)

Restructure each example so nothing mixes: generated code is never in a
hand-written folder, and app never mixes with platform.

  examples/<name>/
    ecu.toml bus.dbc   configuration
    sig/  (module sig)   signal types    ┐ hand (app)
    app/  (module app)   Function Blocks ┘
    main.v (module main) IO/bus bridge    platform (hand)
    ports/ (module ports) In/Out structs ┐ generated
    gen/  (module gen)   codec/tables/glue┘
    Makefile             make all

Each example is freestanding: `cd examples/<name> && make all` generates from its
own ecu.toml (repo tools via REPO=../..) and builds via V's -path (short imports
like `import sig`/`import ports`/`import osal`, not coupled to the example name).
Repo `make example NAME=x` delegates to it. Grouped In/Out kept; the
sig<-ports<-app<-gen chain has no import cycle.

Generators emit module sig/ports/gen with short imports; loom2v imports sig in
the glue only when there are local cells. Warnings eliminated: signed-shift in
candb + generated dbc codec (shift as u64 then cast); unused params in the
backend harness.

minimal restructured to match. AGENTS/README/docs updated. Verified: both
examples build warning-free, run on vcan; lint, demo, v test comm/nm green.


* examples: make [import] dbc path example-relative (fix signal-map provenance)

Addresses Codex P1 on PR #10: sigmap now runs from inside the example, so the
ecu.toml dbc path must be relative to the example dir ('bus.dbc'), not repo-root
('examples/<name>/bus.dbc'). With the old path sigmap couldn't load the DBC and
signal-map.md lost all bus-signal provenance (VehicleSpeed/EngineSpeed showed —).
Restored: VehicleSpeed -> Powertrain.VehicleSpeed 0x100 16|12 x0.1 km/h.


---------

+++ tools/candb/candb.v
@@ -92,7 +92,7 @@ pub fn (s Signal) physical(data []u8) f64 {
 	if s.is_signed && s.length > 0 && s.length < 64 {
 		sign_bit := u64(1) << (s.length - 1)
 		if raw & sign_bit != 0 {
-			v = f64(i64(raw) - (i64(1) << s.length)) // two's-complement negative
+			v = f64(i64(raw) - i64(u64(1) << s.length)) // two's-complement negative
 		}
 	}
 	return v * s.factor + s.offset
@@ -134,7 +134,7 @@ pub fn (s Signal) encode(mut data []u8, phys f64) {
 	// round half away from zero; a bare `+ 0.5` truncates negatives wrongly.
 	mut raw := i64(math.round((phys - s.offset) / s.factor))
 	if raw < 0 {
-		raw += i64(1) << s.length
+		raw += i64(u64(1) << s.length)
 	}
 	mask := if s.length >= 64 { ~u64(0) } else { (u64(1) << s.length) - 1 }
 	s.set_raw(mut data, u64(raw) & mask)
